requirements.txt
anyio==3.6.2
boto3==1.26.97
botocore==1.29.97
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
DateTime==5.1
fastapi==0.95.0
findspark==2.0.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
mysql==0.0.3
mysqlclient==2.1.1
py4j==0.10.9.5
pydantic==1.10.7
PyMySQL==1.0.2
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
zope.interface==6.0
